payz-intern/src/main/java/bccard/payzintern/repository/MemoryMemberRepository.java
package bccard.payzintern.repository;

import bccard.payzintern.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Long, Member> store = new HashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(member.getId(), member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        // 스트림 - 반복문 처럼 도는데 메소드 연결해서 파이프라이닝 할 수 잇음
        // lazy evalutation -> 필요할 때만 연산함
        return store.values().stream()
                // 자바 람다 함수
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
